+++ projects/ui-framework/src/lib/lists/tree-list/tree-list-panel/tree-list-panel.component.scss
@@ -1,4 +1,4 @@
-@import '../../style/common-imports';
+@import '../../../style/common-imports';
 
 :host {
   display: inline-block;
