+++ projects/ui-framework/src/lib/lists/tree-list/tree-list-panel/tree-list-panel.interface.ts
@@ -1,7 +1,7 @@
 import { EventEmitter } from '@angular/core';
-import { PanelDefaultPosVer } from '../../popups/panel/panel.enum';
+import { PanelDefaultPosVer } from '../../../popups/panel/panel.enum';
 import { ConnectedPosition, OverlayRef } from '@angular/cdk/overlay';
-import { TreeListValue } from '../tree-list/tree-list.interface';
+import { TreeListValue } from '../tree-list.interface';
 
 export interface TreeListPanelIO {
   panelPosition: PanelDefaultPosVer | ConnectedPosition[];
