+++ projects/ui-framework/src/lib/lists/tree-list/tree-list-panel/tree-list-panel.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { TreeListPanelComponent } from './tree-list-panel.component';
 import { CommonModule } from '@angular/common';
 import { TreeListModule } from '../tree-list/tree-list.module';
-import { PanelPositionService } from '../../popups/panel/panel-position-service/panel-position.service';
+import { PanelPositionService } from '../../../popups/panel/panel-position-service/panel-position.service';
 import { OverlayModule } from '@angular/cdk/overlay';
-import { ListPanelService } from '../list-panel.service';
+import { ListPanelService } from '../../list-panel.service';
 
 @NgModule({
   declarations: [TreeListPanelComponent],
