+++ projects/ui-framework/src/lib/lists/tree-list/tree-list/tree-list.component.scss
@@ -1,5 +1,5 @@
-@import '../../style/common-imports';
-@import '../../style/icons.mixin';
+@import '../../../style/common-imports';
+@import '../../../style/icons.mixin';
 
 $bhl-item-height: $wrap-height;
 $bhl-item-indent: base8(30);
