+++ projects/ui-framework/src/lib/lists/tree-list/tree-list/tree-list.stories.ts
@@ -1,10 +1,10 @@
 import { storiesOf } from '@storybook/angular';
 import { withKnobs } from '@storybook/addon-knobs/angular';
-import { ComponentGroupType } from '../../consts';
+import { ComponentGroupType } from '../../../consts';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { StoryBookLayoutModule } from '../../story-book-layout/story-book-layout.module';
+import { StoryBookLayoutModule } from '../../../story-book-layout/story-book-layout.module';
 import { TreeListModule } from './tree-list.module';
-import { TreeListStoriesCommonProps } from './tree-list.stories.common';
+import { TreeListStoriesCommonProps } from '../tree-list.stories.common';
 import { boolean } from '@storybook/addon-knobs';
 
 const story = storiesOf(ComponentGroupType.Lists, module).addDecorator(
