+++ projects/ui-framework/src/lib/lists/tree-list/tree-select/tree-select.component.scss
@@ -1,4 +1,4 @@
-@import '../../style/common-imports';
+@import '../../../style/common-imports';
 
 :host ::ng-deep {
   .panel-open.panel-below {
