+++ projects/ui-framework/src/lib/lists/tree-list/tree-select/tree-select.stories.ts
@@ -1,13 +1,13 @@
 import { storiesOf } from '@storybook/angular';
-import { ComponentGroupType } from '../../consts';
+import { ComponentGroupType } from '../../../consts';
 import { withKnobs, text, boolean } from '@storybook/addon-knobs';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { StoryBookLayoutModule } from '../../story-book-layout/story-book-layout.module';
+import { StoryBookLayoutModule } from '../../../story-book-layout/story-book-layout.module';
 import { TreeSelectModule } from './tree-select.module';
-import { mockText } from '../../mock.const';
-import { TreeListStoriesCommonProps } from '../tree-list/tree-list.stories.common';
+import { mockText } from '../../../mock.const';
+import { TreeListStoriesCommonProps } from '../tree-list.stories.common';
 
-import formElemsPropsDoc from '../../form-elements/form-elements.properties.md';
+import formElemsPropsDoc from '../../../form-elements/form-elements.properties.md';
 import { action } from '@storybook/addon-actions';
 
 const story = storiesOf(ComponentGroupType.Lists, module).addDecorator(
