+++ projects/ui-framework/src/public_api.ts
@@ -232,12 +232,12 @@ export {
 // Tree  list
 export * from './lib/lists/tree-list/tree-list.interface';
 export * from './lib/lists/tree-list/tree-list.const';
-export { TreeListModule } from './lib/lists/tree-list/tree-list.module';
-export { TreeListComponent } from './lib/lists/tree-list/tree-list.component';
-export {TreeListPanelModule}from './lib/lists/tree-list-panel/tree-list-panel.module';
-export { TreeListPanelComponent } from './lib/lists/tree-list-panel/tree-list-panel.component';
-export { TreeSelectModule } from './lib/lists/tree-select/tree-select.module';
-export { TreeSelectComponent } from './lib/lists/tree-select/tree-select.component';
+export { TreeListModule } from './lib/lists/tree-list/tree-list/tree-list.module';
+export { TreeListComponent } from './lib/lists/tree-list/tree-list/tree-list.component';
+export {TreeListPanelModule}from './lib/lists/tree-list/tree-list-panel/tree-list-panel.module';
+export { TreeListPanelComponent } from './lib/lists/tree-list/tree-list-panel/tree-list-panel.component';
+export { TreeSelectModule } from './lib/lists/tree-list/tree-select/tree-select.module';
+export { TreeSelectComponent } from './lib/lists/tree-list/tree-select/tree-select.component';
 export { TreeListModelService } from './lib/lists/tree-list/services/tree-list-model.service';
 
 // Chain select
